moved parts of the connect logic into lib
