+++ lib/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod import;
 pub mod clean;
+pub mod connect;
